front: add diagram to switch types editor
add node type diagram to node form
