+++ front/src/applications/editor/tools/switchEdition/components.tsx
@@ -1,223 +1,54 @@
-import React, { FC, useCallback, useContext, useEffect, useMemo, useState } from 'react';
-import { FieldProps } from '@rjsf/core';
+import React, { FC, useContext, useEffect, useMemo, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 import Select from 'react-select';
-import { first, last, keyBy, debounce } from 'lodash';
-import { FaTimesCircle, FaMapMarkedAlt } from 'react-icons/fa';
-import SchemaField from '@rjsf/core/lib/components/fields/SchemaField';
+import { first, last, debounce } from 'lodash';
 import { Layer, Popup, Source } from 'react-map-gl/maplibre';
 import { featureCollection, point } from '@turf/helpers';
 import nearestPoint from '@turf/nearest-point';
 import { Position } from 'geojson';
 
 import EditorContext from '../../context';
-import Tipped from '../../components/Tipped';
-import {
-  EntityObjectOperationResult,
-  DEFAULT_ENDPOINT,
-  ENDPOINTS,
-  ENDPOINTS_SET,
-  SwitchEntity,
-  TrackSectionEntity,
-} from '../../../../types';
+import { EntityObjectOperationResult, SwitchEntity, TrackSectionEntity } from '../../../../types';
 import colors from '../../../../common/Map/Consts/colors';
 import GeoJSONs from '../../../../common/Map/Layers/GeoJSONs';
-import { PortEndPointCandidate, SwitchEditionState, useSwitchTypes } from './types';
+import { SwitchEditionState } from './types';
 import EditorForm from '../../components/EditorForm';
 import { save } from '../../../../reducers/editor';
-import {
-  FLAT_SWITCH_PORTS_PREFIX,
-  FlatSwitchEntity,
-  flatSwitchToSwitch,
-  getNewSwitch,
-  getSwitchTypeJSONSchema,
-  isSwitchValid,
-  switchToFlatSwitch,
-} from './utils';
+import { FlatSwitchEntity, flatSwitchToSwitch, getNewSwitch, isSwitchValid } from './utils';
 import {
   getSwitchesLayerProps,
   getSwitchesNameLayerProps,
 } from '../../../../common/Map/Layers/Switches';
-import { flattenEntity, NEW_ENTITY_ID } from '../../data/utils';
+import { flattenEntity } from '../../data/utils';
 import EntitySumUp from '../../components/EntitySumUp';
 import EntityError from '../../components/EntityError';
 import { getEntity } from '../../data/api';
 import { getInfraID } from '../../../../reducers/osrdconf/selectors';
 import { getMap } from '../../../../reducers/map/selectors';
 import { ExtendedEditorContextType } from '../editorContextTypes';
-
-const ENDPOINT_OPTIONS = ENDPOINTS.map((s) => ({ value: s, label: s }));
-const ENDPOINT_OPTIONS_DICT = keyBy(ENDPOINT_OPTIONS, 'value');
-
-export const TrackSectionEndpointSelector: FC<FieldProps> = ({
-  schema,
-  formData,
-  onChange,
-  name,
-}) => {
-  const dispatch = useDispatch();
-  const { state, setState } = useContext(
-    EditorContext
-  ) as ExtendedEditorContextType<SwitchEditionState>;
-  const { t } = useTranslation();
-  const infraID = useSelector(getInfraID);
-
-  const portId = name.replace(FLAT_SWITCH_PORTS_PREFIX, '');
-  const endpoint = ENDPOINTS_SET.has(formData?.endpoint) ? formData.endpoint : DEFAULT_ENDPOINT;
-  const [trackSection, setTrackSection] = useState<TrackSectionEntity | null>(null);
-
-  const isPicking =
-    state.portEditionState.type === 'selection' && state.portEditionState.portId === portId;
-  const isDisabled =
-    state.portEditionState.type === 'selection' && state.portEditionState.portId !== portId;
-
-  const startPickingPort = useCallback(() => {
-    // Cancel current selection:
-    if (isPicking) {
-      setState({
-        ...state,
-        portEditionState: {
-          type: 'idle',
-        },
-      });
-    }
-    // Start selecting:
-    else if (!isDisabled) {
-      setState({
-        ...state,
-        portEditionState: {
-          type: 'selection',
-          portId,
-          hoveredPoint: null,
-          onSelect: (hoveredPoint: PortEndPointCandidate) => {
-            onChange({
-              endpoint: hoveredPoint.endPoint,
-              track: hoveredPoint.trackSectionId,
-            });
-          },
-        },
-      });
-    }
-  }, [isDisabled, isPicking, onChange, portId, setState, state]);
-
-  useEffect(() => {
-    if (typeof formData?.track === 'string') {
-      getEntity<TrackSectionEntity>(
-        infraID as number,
-        formData.track,
-        'TrackSection',
-        dispatch
-      ).then((track) => {
-        setTrackSection(track);
-      });
-    } else {
-      setTrackSection(null);
-    }
-  }, [formData?.track, infraID]);
-
-  return (
-    <div className="mb-4">
-      {schema.title && <h5>{schema.title}</h5>}
-      {schema.description && <p>{schema.description}</p>}
-      <div className="d-flex flex-row align-items-center mb-2">
-        <div className="flex-grow-1 flex-shrink-1 mr-2">
-          {trackSection ? (
-            <span>
-              {trackSection?.properties?.extensions?.sncf?.line_name || trackSection.properties.id}
-            </span>
-          ) : (
-            <span className="text-danger font-weight-bold">
-              {t('Editor.tools.switch-edition.no-track-picked-yet')}
-            </span>
-          )}
-          {!!trackSection && <div className="text-muted small">{trackSection.properties.id}</div>}
-          {!!trackSection && (
-            <div className="d-flex flex-row align-items-baseline mb-2">
-              <span className="mr-2">{t('Editor.tools.switch-edition.endpoint')}</span>
-              <Select
-                value={ENDPOINT_OPTIONS_DICT[endpoint]}
-                options={ENDPOINT_OPTIONS}
-                onChange={(o) => {
-                  if (o)
-                    onChange({
-                      endpoint: o.value,
-                      track: trackSection.properties.id,
-                    });
-                }}
-              />
-            </div>
-          )}
-        </div>
-        <Tipped mode="left">
-          <button
-            type="button"
-            className="btn btn-primary px-3"
-            onClick={startPickingPort}
-            disabled={isDisabled}
-          >
-            {isPicking ? <FaTimesCircle /> : <FaMapMarkedAlt />}
-          </button>
-          <span>
-            {t(
-              `Editor.tools.switch-edition.actions.${
-                isPicking ? 'pick-track-cancel' : 'pick-track'
-              }`
-            )}
-          </span>
-        </Tipped>
-      </div>
-    </div>
-  );
-};
-
-export const CustomSchemaField: FC<FieldProps> = (props) => {
-  const { name = '' } = props;
-  if (name.indexOf(FLAT_SWITCH_PORTS_PREFIX) !== 0) return <SchemaField {...props} />;
-  return <TrackSectionEndpointSelector {...props} />;
-};
+import { CustomSchemaField } from './components/CustomSchemaField';
+import useSwitch from './useSwitch';
 
 export const SwitchEditionLeftPanel: FC = () => {
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const infraID = useSelector(getInfraID);
-  const { state, setState, editorState } = useContext(
+  const { state, setState } = useContext(
     EditorContext
   ) as ExtendedEditorContextType<SwitchEditionState>;
 
-  // Retrieve base JSON schema:
-  const baseSchema = editorState.editorSchema.find((e) => e.objType === 'Switch')?.schema;
   // Retrieve proper data
-  const switchTypes = useSwitchTypes(infraID);
-  const switchTypesDict = useMemo(() => keyBy(switchTypes, 'id'), [switchTypes]);
-  const switchTypeOptions = useMemo(
-    () =>
-      (switchTypes || []).map((type) => ({
-        value: type.id,
-        label: `${type.id} (${type.ports.length} port${type.ports.length > 1 ? 's' : ''})`,
-      })),
-    [switchTypes]
-  );
-  const switchTypeOptionsDict = useMemo(
-    () => keyBy(switchTypeOptions, 'value'),
-    [switchTypeOptions]
-  );
-  const switchEntity = state.entity as SwitchEntity;
-  const isNew = switchEntity.properties.id === NEW_ENTITY_ID;
-  const switchType = useMemo(
-    () =>
-      switchTypes?.find((type) => type.id === switchEntity.properties.switch_type) ||
-      first(switchTypes || []),
-    [switchEntity.properties.switch_type, switchTypes]
-  );
-  const flatSwitchEntity = useMemo(
-    () => switchType && switchToFlatSwitch(switchType, switchEntity),
-    [switchEntity, switchType]
-  );
-  const switchTypeJSONSchema = useMemo(
-    () => switchType && baseSchema && getSwitchTypeJSONSchema(baseSchema, switchType),
-    [baseSchema, switchType]
-  );
+  const {
+    switchEntity,
+    flatSwitchEntity,
+    switchType,
+    switchTypeOptions,
+    switchTypeOptionsDict,
+    switchTypesDict,
+    switchTypeJSONSchema,
+    isNew,
+  } = useSwitch();
 
   if (!switchType || !flatSwitchEntity) {
     return null;
@@ -311,7 +142,6 @@ export const SwitchEditionLayers: FC = () => {
   const dispatch = useDispatch();
   const { t } = useTranslation();
   const infraID = useSelector(getInfraID);
-  const switchTypes = useSwitchTypes(infraID);
   const {
     renderingFingerprint,
     state,
@@ -319,11 +149,8 @@ export const SwitchEditionLayers: FC = () => {
     editorState: { editorLayers },
   } = useContext(EditorContext) as ExtendedEditorContextType<SwitchEditionState>;
   const { entity, hovered, portEditionState, mousePosition } = state;
-  const switchTypesDict = useMemo(() => keyBy(switchTypes, 'id'), [switchTypes]);
-  const switchType = useMemo(
-    () => (entity.properties?.switch_type ? switchTypesDict[entity.properties.switch_type] : null),
-    [entity.properties?.switch_type, switchTypesDict]
-  );
+
+  const { switchType } = useSwitch();
   const { mapStyle, layersSettings, issuesSettings } = useSelector(getMap);
   const layerProps = useMemo(
     () =>
